+++ compiler/compiler-plugin/src/testFixtures/kotlin/com/cmgapps/logtag/compiler/runner/AbstractJvmBoxTest.kt
@@ -6,7 +6,9 @@
 
 package com.cmgapps.logtag.compiler.runner
 
+import com.cmgapps.logtag.LogTagConfigurationKeys
 import com.cmgapps.logtag.compiler.service.configurePlugin
+import org.jetbrains.kotlin.config.CompilerConfigurationKey
 import org.jetbrains.kotlin.test.builders.TestConfigurationBuilder
 import org.jetbrains.kotlin.test.directives.CodegenTestDirectives
 import org.jetbrains.kotlin.test.directives.FirDiagnosticsDirectives
@@ -18,6 +20,11 @@ import org.jetbrains.kotlin.test.services.KotlinStandardLibrariesPathProvider
 open class AbstractJvmBoxTest : AbstractFirBlackBoxCodegenTestSpec() {
     override fun createKotlinStandardLibrariesPathProvider(): KotlinStandardLibrariesPathProvider =
         EnvironmentBasedStandardLibrariesPathProvider
+
+    open val configurationMap: Map<CompilerConfigurationKey<*>, Any> =
+        mapOf(
+            LogTagConfigurationKeys.ENABLED to true,
+        )
 
     override fun configure(builder: TestConfigurationBuilder) =
         with(builder) {
@@ -41,6 +48,6 @@ open class AbstractJvmBoxTest : AbstractFirBlackBoxCodegenTestSpec() {
                 +CodegenTestDirectives.IGNORE_DEXING // Avoids loading R8 from the classpath.
             }
 
-            configurePlugin()
+            configurePlugin(configurationMap = configurationMap)
         }
 }