+++ compiler/gradle-plugin/src/main/kotlin/com/cmgapps/gradle/LogTagGradleSupportPlugin.kt
@@ -6,31 +6,61 @@
 
 package com.cmgapps.gradle
 
+import com.android.build.api.dsl.CommonExtension
+import com.android.build.api.dsl.KotlinMultiplatformAndroidLibraryExtension
 import com.cmgapps.logtag.gradle.BuildConfig
 import org.gradle.api.Project
-import org.gradle.api.provider.Property
 import org.gradle.api.provider.Provider
+import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
 import org.jetbrains.kotlin.gradle.plugin.KotlinCompilation
 import org.jetbrains.kotlin.gradle.plugin.KotlinCompilerPluginSupportPlugin
 import org.jetbrains.kotlin.gradle.plugin.SubpluginArtifact
 import org.jetbrains.kotlin.gradle.plugin.SubpluginOption
 
-@Suppress("kotlin:S6526")
-public abstract class LogTagExtension {
-    public abstract val enabled: Property<Boolean>
-
-    init {
-        enabled.convention(true)
-    }
-}
+private val ANDROID_IDS =
+    listOf(
+        "com.android.application",
+        "com.android.library",
+        "com.android.feature",
+        "com.android.dynamic-feature",
+        "com.android.test",
+    )
 
 @Suppress("unused")
 public class LogTagGradleSupportPlugin : KotlinCompilerPluginSupportPlugin {
     override fun apply(target: Project) {
         with(target) {
-            extensions.create("logTag", LogTagExtension::class.java)
+            val logTagExtension = objects.newInstance(LogTagExtensionInternal::class.java)
+            logTagExtension.minSdk.convention(Int.MAX_VALUE)
+            extensions.add(LogTagExtension::class.java, "logTag", logTagExtension)
 
-            // TODO: check if android project and apply the linter library
+            ANDROID_IDS.forEach {
+                plugins.withId(it) {
+                    dependencies.add("implementation", BuildConfig.ANDROID_LINT_LIBRARY_COORDINATES)
+
+                    val commonAndroidExtension = extensions.getByType(CommonExtension::class.java)
+                    afterEvaluate {
+                        logTagExtension.minSdk.set(commonAndroidExtension.defaultConfig.minSdk)
+                    }
+                }
+            }
+
+            plugins.withId("com.android.kotlin.multiplatform.library") {
+                val multiplatformExtension = extensions.getByType(KotlinMultiplatformExtension::class.java)
+                multiplatformExtension.sourceSets.named("androidMain") {
+                    it.dependencies {
+                        implementation(BuildConfig.ANDROID_LINT_LIBRARY_COORDINATES)
+                    }
+                }
+
+                multiplatformExtension.extensions
+                    .findByType(KotlinMultiplatformAndroidLibraryExtension::class.java)
+                    ?.let { androidExtension ->
+                        afterEvaluate {
+                            logTagExtension.minSdk.set(androidExtension.minSdk)
+                        }
+                    }
+            }
         }
     }
 
@@ -49,11 +79,10 @@ public class LogTagGradleSupportPlugin : KotlinCompilerPluginSupportPlugin {
         }
 
         return with(kotlinCompilation.target.project) {
-            val extension = extensions.getByType(LogTagExtension::class.java)
+            val extension = extensions.getByType(LogTagExtension::class.java) as LogTagExtensionInternal
             objects.listProperty(SubpluginOption::class.java).apply {
-                add(
-                    extension.enabled.map { SubpluginOption("enabled", it.toString()) },
-                )
+                add(extension.enabled.map { SubpluginOption("enabled", it.toString()) })
+                add(extension.minSdk.map { SubpluginOption("android.minSdk", it.toString()) })
             }
         }
     }