+++ example-compiler-plugin/src/commonMain/kotlin/foo/bar/Foo.kt
@@ -1,17 +1,7 @@
 /*
  * Copyright (c) 2021. Christian Grach <christian.grach@cmgapps.com>
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * SPDX-License-Identifier: Apache-2.0
  */
 
 package foo.bar
@@ -95,7 +85,7 @@ value class ValueClass(
     override fun log(message: String) = println("$LOG_TAG:$value -> $message")
 }
 
-@androidx.compose.runtime.Composable
+@Suppress("ktlint:standard:function-naming")
 @LogTag
 fun Test() {
     println("$LOG_TAG -> annotated function")
@@ -116,25 +106,7 @@ fun main() {
         ClassWithCompanion(),
         ValueClass(42),
     ).forEach { it.log("Hello, World!") }
-
     Works.entries.forEach { it.log() }
 
     Test()
-    println(box())
-}
-
-fun localTag(): String {
-    @LogTag
-    fun tagged(): String = LOG_TAG
-
-    return tagged()
-}
-
-fun box(): String {
-    val result = localTag()
-    return if (result == "tagged") {
-        "OK"
-    } else {
-        "FAIL"
-    }
 }