+++ processor/src/functionalTest/resources/test/create_composable_tag/class.kt
@@ -1,0 +1,5 @@
+package cmgapps.test
+
+@com.cmgapps.LogTag
+@androidx.compose.runtime.Composable
+fun Test() {}