+++ processor/src/functionalTest/resources/test/create_composable_tag/expected/kotlin/cmgapps/test/ComposableTest.kt
@@ -1,0 +1,16 @@
+@file:Suppress(
+  "SpellCheckingInspection",
+  "RedundantVisibilityModifier",
+  "unused",
+)
+
+package cmgapps.test
+
+import kotlin.String
+import kotlin.Suppress
+
+public class ComposableTest {
+  public companion object {
+    public const val LOG_TAG: String = "Test"
+  }
+}