+++ processor/src/functionalTest/resources/test/generate_class_for_java_class/TestJava.java
@@ -1,0 +1,4 @@
+package cmgapps.test;
+
+@com.cmgapps.LogTag
+public class TestJava {}