+++ processor/src/functionalTest/resources/test/generate_extension_for_kotlin_class/class.kt
@@ -1,0 +1,10 @@
+/*
+ * Copyright (c) 2026. Christian Grach <christian.grach@cmgapps.com>
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package cmgapps.test
+
+@com.cmgapps.LogTag
+class TestClass