+++ processor/src/functionalTest/resources/test/generate_for_internal_class/class.kt
@@ -1,0 +1,4 @@
+package cmgapps.test
+
+@com.cmgapps.LogTag
+internal class TestClass