+++ processor/src/functionalTest/resources/test/generate_for_internal_class/expected/kotlin/cmgapps/test/TestClassLogTag.kt
@@ -1,0 +1,13 @@
+@file:Suppress(
+  "SpellCheckingInspection",
+  "RedundantVisibilityModifier",
+  "unused",
+)
+
+package cmgapps.test
+
+import kotlin.String
+import kotlin.Suppress
+
+internal inline val TestClass.LOG_TAG: String
+  get() = "TestClass"