+++ processor/src/functionalTest/resources/test/use_custom_tag_for_java_class/TestJava.java
@@ -1,0 +1,4 @@
+package cmgapps.test;
+
+@com.cmgapps.LogTag("MyCustomTag")
+public class TestJava {}