+++ processor/src/functionalTest/resources/test/use_custom_tag_for_java_class/expected/java/cmgapps/test/TestJavaLogTag.java
@@ -1,0 +1,7 @@
+package cmgapps.test;
+
+import java.lang.String;
+
+class TestJavaLogTag {
+  static final String LOG_TAG = "MyCustomTag";
+}